requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.2
ccxt==1.66.17
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pkg_resources==0.0.0
pycares==4.1.2
pycparser==2.21
python-dateutil==2.8.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarl==1.7.2
